+++ packages/charts/src/charts/pie/pie-chart.component.tsx
@@ -8,7 +8,7 @@ import { getTooltipContent } from '../../utils/tooltip.utils';
 import PieSlice from './pie-slice.component';
 import ShadowFilter from '../../components/shadow-filter.component';
 
-import { ChartBase } from '../../components';
+import { ChartBase, Delayed } from '../../components';
 
 import { useTooltip } from '../../hooks';
 import { theme as defaultTheme } from '../../theme';
@@ -126,54 +126,58 @@ export const PieChart: FC<Props> = ({
           </motion.div>
         )}
       </AnimatePresence>
-      <ChartBase
-        ref={svgElement}
-        theme={theme}
-        svgDimensions={svgDimensions}
-        margins={margins}
-      >
-        <g
-          style={{
-            transform: `translate(${svgDimensions.width /
-              2}px, ${svgDimensions.height / 2}px)`,
-          }}
+      <Delayed>
+        <ChartBase
+          ref={svgElement}
+          theme={theme}
+          svgDimensions={svgDimensions}
+          margins={margins}
         >
-          <ShadowFilter />
-          {arcs.map(
-            ({
-              index,
-              label,
-              labelPosition,
-              activePosition,
-              startAngle,
-              endAngle,
-              color,
-              selector,
-              stacked,
-              stack,
-            }) => (
-              <PieSlice
-                key={index}
-                draw={drawArc}
-                startAngle={startAngle}
-                endAngle={endAngle}
-                label={label}
-                autocolor={labelsAutocolor}
-                activePosition={activePosition}
-                labelPosition={labelPosition}
-                background={color}
-                onMouseMove={e => {
-                  if (tooltipSettings.enabled) {
-                    if (stacked) updateTooltipPosition(e, stack);
-                    else updateTooltipPosition(e, [{ color, selector }]);
-                  }
-                }}
-                onMouseLeave={() => hideTooltip()}
-              />
-            )
-          )}
-        </g>
-      </ChartBase>
+          <g
+            style={{
+              transform: `translate(${svgDimensions.width /
+                2}px, ${svgDimensions.height / 2}px)`,
+            }}
+          >
+            <ShadowFilter />
+            <AnimatePresence>
+              {arcs.map(
+                ({
+                  label,
+                  labelPosition,
+                  activePosition,
+                  dataKey,
+                  startAngle,
+                  endAngle,
+                  color,
+                  selector,
+                  stacked,
+                  stack,
+                }) => (
+                  <PieSlice
+                    key={dataKey}
+                    draw={drawArc}
+                    startAngle={startAngle}
+                    endAngle={endAngle}
+                    label={label}
+                    autocolor={labelsAutocolor}
+                    activePosition={activePosition}
+                    labelPosition={labelPosition}
+                    background={color}
+                    onMouseMove={e => {
+                      if (tooltipSettings.enabled) {
+                        if (stacked) updateTooltipPosition(e, stack);
+                        else updateTooltipPosition(e, [{ color, selector }]);
+                      }
+                    }}
+                    onMouseLeave={() => hideTooltip()}
+                  />
+                )
+              )}
+            </AnimatePresence>
+          </g>
+        </ChartBase>
+      </Delayed>
     </>
   );
 };