+++ packages/charts/src/components/pie-label/pie-label.component.tsx
@@ -3,21 +3,13 @@ import React, { FC, memo } from 'react';
 import { ColorAdjuster, Typography } from '@keen.io/ui-core';
 
 type Props = {
-  position: [number, number];
   children: React.ReactNode;
   sliceBackground: string;
   autocolor: boolean;
 } & Typography;
 
 const PieLabel: FC<Props> = memo(
-  ({
-    position,
-    sliceBackground,
-    autocolor,
-    children,
-    fontColor,
-    ...typography
-  }) => {
+  ({ sliceBackground, autocolor, children, fontColor, ...typography }) => {
     return (
       <>
         {autocolor ? (
@@ -26,7 +18,6 @@ const PieLabel: FC<Props> = memo(
               <text
                 fill={adjustedColor}
                 pointerEvents="none"
-                transform={`translate(${position})`}
                 style={{ textAnchor: 'middle', ...typography }}
               >
                 {children}
@@ -37,7 +28,6 @@ const PieLabel: FC<Props> = memo(
           <text
             fill={fontColor}
             pointerEvents="none"
-            transform={`translate(${position})`}
             style={{ textAnchor: 'middle', ...typography }}
           >
             {children}