+++ packages/icons/src/components/utils.ts
@@ -8,6 +8,7 @@ import Brand from './brand.component';
 import QuestionMark from './question-mark.component';
 import EyeSolid from './eye-solid.component';
 import EyeOutline from './eye-outline.component';
+import EyeSlashSolid from './eye-slash-solid.component';
 import UserSolid from './user-solid.component';
 import UserOutline from './user-outline.component';
 import UsersSolid from './users-solid.component';
@@ -16,6 +17,8 @@ import ClickSolid from './click-solid.component';
 import ClickOutline from './click-outline.component';
 import CursorSolid from './cursor-solid.component';
 import CursorOutline from './cursor-outline.component';
+import Check from './check.component';
+import Line from './line.component';
 
 import { IconType } from '../types';
 
@@ -33,6 +36,10 @@ export const getIcon = (type: IconType) => {
       return CaretLeft;
     case 'caret-right':
       return CaretRight;
+    case 'check':
+      return Check;
+    case 'line':
+      return Line;
     case 'brand':
       return Brand;
     case 'question-mark':
@@ -49,6 +56,8 @@ export const getIcon = (type: IconType) => {
       return EyeSolid;
     case 'eye-outline':
       return EyeOutline;
+    case 'eye-slash-solid':
+      return EyeSlashSolid;
     case 'user-solid':
       return UserSolid;
     case 'user-outline':