+++ packages/ui-core/src/components/error/index.ts
@@ -1,3 +1,3 @@
 import Error from './error.component';
 
-export { Error };
+export default Error;