+++ packages/ui-core/src/components/password-input/index.ts
@@ -1,0 +1,3 @@
+import PasswordInput from './password-input.component';
+
+export default PasswordInput;