+++ packages/ui-core/src/components/password-input/password-input.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+
+import { PasswordInput } from './password-input.component';
+
+export default {
+  title: 'Components| Forms / Password Input',
+  parameters: {
+    component: PasswordInput,
+    componentSubtitle: 'Input dedicated for password',
+  },
+};
+
+export const basic = () => (
+  <div style={{ width: '250px' }}>
+    <PasswordInput defaultValue="keen.io" />
+  </div>
+);